Fixed inconsistency that caused IndexOutOfBounds
Whenever a settlement is not built, its ownerId is -1.  The game
manager called “Player p = TurnState.players[settlement.getPlayer()];”
This line would be executed always, because the check “if
(settlement.city)” was always true, because the city was never null.
Since settlement.ownerID was -1, this would cause an Exception.  Added
“hasCity” bool and “isCity” function to SettlementClass to help solve
this, so it only goes through that code path when it actually should.