Added ability to 4:1 trade during trade phase
A console will pop up during the trade phase, allowing you to trade 4
of 1 resource for 1 of another, with the house. **Current issue**: It
pops up right over the dev console; not sure where the best place to
put it is.  Since dev console is only used in development, you can
disable the TradeConsole through the Scene view if you need to use the
DevConsole during the trade phase.