Made TradeConsole reset after trades
Now, after trading, it resets.  This allows it to account for changes
made in the trade.  For example, if you have 3 wheat and 4 brick, then
trade the 4 brick for a wheat, you have 4 wheat, which is enough to
trade.  Before, it wouldn’t let you make this trade.  Now, it does.