Put placeholders for RPC calls
Modified the Settlement, Road, and Tile classes so that on click events
are split into two cases: client will send request to server to perform
the desired action, or the server will perform the action itself and
update the clients.

Still need to go through other sections of the code and find where
other events require server intervention.