Turns are now working
Turns work, moving the robber steals correctly. One thing we need to
change with the robber is: when the dev card is played, the TurnState
is changed to robbering, and the same should happen when a 7 is rolled.
We should not allow any other buttons to be pressed during this time (I
did this), other than selecting a tile and possibly a settlement if
there is one. We also have to exit the state of robbering, which could
be either trade-phase or build-phase (I think). We need to
differentiate between the two cases.