Game start now works
Given that a particular button is pressed (this can be changed later),
the networking required to keep the players in sync and play through
the setup of the game is now working. Now to fix the robber issue and
then merge with master once Charlie has the UI’s for the
lobby/creating/joining games.